fix test

+++ internal/infrastructure/db/influx/market_price_repository_impl.go
@@ -58,9 +58,9 @@ func (i *influxDbService) GetPricesForMarkets(
 		marketID := result.Record().ValueByKey(marketTag).(string)
 		marketPrice := domain.MarketPrice{
 			MarketID:   result.Record().ValueByKey(marketTag).(string),
-			BasePrice:  result.Record().ValueByKey(basePrice).(float32),
+			BasePrice:  float32(result.Record().ValueByKey(basePrice).(float64)),
 			BaseAsset:  result.Record().ValueByKey(baseAsset).(string),
-			QuotePrice: result.Record().ValueByKey(basePrice).(float32),
+			QuotePrice: float32(result.Record().ValueByKey(basePrice).(float64)),
 			QuoteAsset: result.Record().ValueByKey(quoteAsset).(string),
 			Time:       result.Record().Time(),
 		}
